+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The main work (all changes without a GitHub username in brackets in the below li
 - @matter/protocol
     - Fix: Handles messages only that are secured as required for the relevant protocol
 
+- @matter/general
+    - Enhancement: Allows async implementation of crypto methods
+
 ## 0.13.0 (2025-04-28)
 
 -   IMPORTANT: This release upgrades Matter support from Matter 1.3 to the latest release, Matter 1.4.0. This includes BREAKING CHANGES in a number of areas due to specification changes. For the most part these changes are transparent because they involve low-level APIs, implicit type names, or Matter features that were never adopted elsewhere. However, some small code changes may be necessary depending on how you use Matter.js or when Datatypes or elements got renamed.